Cantidad carrito

+++ src/components/views/pedidos/Pedidos.js
@@ -5,39 +5,76 @@ import Swal from "sweetalert2";
 import ItemPedidos from "./ItemPedidos";
 import { useNavigate } from "react-router-dom";
 import "./Pedidos.css";
+import { formatMoneda } from "../helperCarrito";
 
 const Pedidos = () => {
     const navigate = useNavigate();
-    let total = 0;
     const URL = process.env.REACT_APP_API_HAMBURGUESERIA;
-
+    
     useEffect(() => {
+        actualizarTotal(listaProductosPedido);
         window.scrollTo(0, 0);
     }, []);
-
-    const productosPedido = JSON.parse(localStorage.getItem(process.env.REACT_APP_LOCALSTORAGE_PRODUCTOS_PEDIDO)) || [];
+    
+    const productosPedidoTemp = JSON.parse(localStorage.getItem(process.env.REACT_APP_LOCALSTORAGE_PRODUCTOS_PEDIDO)) || [];
     const usuario = JSON.parse(localStorage.getItem(process.env.REACT_APP_LOCALSTORAGE)) || { nombre: "anonimo!!" };
-    const [listaProductosPedido, setListaProductosPedido] = useState(productosPedido);
+    const [listaProductosPedido, setListaProductosPedido] = useState(productosPedidoTemp);
+    const [total, setTotal] = useState(0);
+
+    const actualizarTotal=(lista)=>{
+        let subTotal = 0;
+        lista.map((item)=>subTotal += item.precio * item.cantidad);
+        setTotal(subTotal);
+    }
 
     const quitarProducto = (producto) => {
-        let nuevaLista = listaProductosPedido.filter((item) => {
-            return item._id !== producto._id;
+        Swal.fire({
+            title: "Esta seguro?",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Si, eliminar",
+            cancelButtonText: "Cancelar",
+        }).then((result) => {
+            if (result.isConfirmed) {
+                let nuevaLista = listaProductosPedido.filter((item) => {
+                    return item._id !== producto._id;
+                });
+                actualizarTotal(nuevaLista);
+                setListaProductosPedido(nuevaLista);
+                localStorage.setItem(process.env.REACT_APP_LOCALSTORAGE_PRODUCTOS_PEDIDO, JSON.stringify(nuevaLista));
+                Swal.fire("Producto eliminado", "El producto fue quitado del pedido", "success");
+            }
         });
-        setListaProductosPedido(nuevaLista);
-        localStorage.setItem(process.env.REACT_APP_LOCALSTORAGE_PRODUCTOS_PEDIDO, JSON.stringify(nuevaLista));
-        Swal.fire("Producto eliminado", "El producto fue quitado del pedido", "success");
     };
+
+    const restarUno = (producto) => {
+        let i = listaProductosPedido.findIndex((item) => {
+            return item._id === producto._id;
+        });
+        if (i>=0 && listaProductosPedido[i].cantidad>1) 
+        {
+            listaProductosPedido[i].cantidad--;
+            actualizarTotal(listaProductosPedido);
+            localStorage.setItem(process.env.REACT_APP_LOCALSTORAGE_PRODUCTOS_PEDIDO, JSON.stringify(listaProductosPedido));
+        }
+    }
+
+    const sumarUno = (producto) => {
+        let i = listaProductosPedido.findIndex((item) => {
+            return item._id === producto._id;
+        });
+        if (i>=0) 
+        {
+            listaProductosPedido[i].cantidad++;
+            actualizarTotal(listaProductosPedido);
+            localStorage.setItem(process.env.REACT_APP_LOCALSTORAGE_PRODUCTOS_PEDIDO, JSON.stringify(listaProductosPedido));
+        }
+    }
 
     const guardarPedido = async () => {
         try {
-            let productosPedido = [];
-            total = 0;
-            listaProductosPedido.forEach((element) => {
-                total += element.precio;
-                element.cantidad = 1;
-                productosPedido.push(element);
-            });
-
             const today = new Date();
             let day = today.getDate();
             if (day < 10) day = "0" + day;
@@ -50,7 +87,7 @@ const Pedidos = () => {
             const pedidos = {
                 usuario: usuario.nombre,
                 fecha,
-                productosdelmenu: [...productosPedido],
+                productosdelmenu: [...listaProductosPedido],
                 estado: false, // false=PENDIENTE
             };
             console.log(pedidos);
@@ -90,10 +127,7 @@ const Pedidos = () => {
     }
 
     const handleClick = () => {
-        total = 0;
-        listaProductosPedido.forEach((element) => {
-            total += element.precio;
-        });
+        actualizarTotal(listaProductosPedido);
 
         Swal.fire({
             title: "Esta seguro?",
@@ -118,19 +152,20 @@ const Pedidos = () => {
                 <Table striped responsive>
                     <thead>
                         <tr>
-                            <th>Item</th>
                             <th>Producto</th>
                             <th>Cantidad</th>
                             <th>Precio</th>
+                            <th>Sub-total</th>
                             <th>Accion</th>
                         </tr>
                     </thead>
                     <tbody>
                         {listaProductosPedido.map((producto) => (
-                            <ItemPedidos key={producto._id} producto={producto} quitarProducto={quitarProducto}></ItemPedidos>
+                            <ItemPedidos key={producto._id} producto={producto} quitarProducto={quitarProducto} restarUno={restarUno} sumarUno={sumarUno}></ItemPedidos>
                         ))}
                     </tbody>
                 </Table>
+                <p>Total: {formatMoneda(total)}</p>
                 <div className="text-end">
                     <Button variant="danger" className="mt-3 me-3 text-light" onClick={borrarCarrito}>Borrar carrito</Button>
                     <Button variant="primary" className="mt-3" onClick={handleClick}>
